refactor: update engine and input handling for improved precision and clarity; add asset management functionality

+++ src/engine.cpp
@@ -86,7 +86,7 @@ namespace RealmEngine
     void Engine::tick()
     {
         double current_time = glfwGetTime();
-        m_delta_time        = static_cast<float>(current_time - m_last_frame_time);
+        m_delta_time        = current_time - m_last_frame_time;
         m_last_frame_time   = current_time;
         if (m_delta_time > m_max_delta_time)
             m_delta_time = m_max_delta_time;
